[NX-OS] Implement support for `Aggregate` interfaces on Cisco NX-OS
This change implements the `type:Aggregate` for the `Interface`
resources as `Port-channel` interfaces on Cisco NX-OS.

+++ internal/provider/cisco/nxos/provider.go
@@ -481,7 +481,7 @@ func (p *Provider) EnsureInterface(ctx context.Context, req *provider.InterfaceR
 		}
 	}
 
-	if req.IPv4 == nil {
+	if req.Interface.Spec.Type != v1alpha1.InterfaceTypeAggregate && req.IPv4 == nil {
 		// Ensure to delete any leftover IPv4 addresses if the spec does not contain any.
 		if err := p.client.Delete(ctx, addr); err != nil {
 			return err
@@ -545,8 +545,7 @@ func (p *Provider) EnsureInterface(ctx context.Context, req *provider.InterfaceR
 		lb.RtvrfMbrItems = NewVrfMember(name, DefaultVRFName)
 		conf = append(conf, lb)
 
-	// case v1alpha1.InterfaceTypePortChannel:
-	case "PortChannel":
+	case v1alpha1.InterfaceTypeAggregate:
 		f := new(Feature)
 		f.Name = "lacp"
 		f.AdminSt = AdminStEnabled
@@ -563,17 +562,34 @@ func (p *Provider) EnsureInterface(ctx context.Context, req *provider.InterfaceR
 		pc := new(PortChannel)
 		pc.ID = name
 		pc.Descr = req.Interface.Spec.Description
-		pc.PcMode = PortChannelModeActive
-		pc.AccessVlan = DefaultVLAN
 		pc.AdminSt = AdminStDown
-		pc.Layer = Layer2
-		pc.Mode = SwitchportModeAccess
-		pc.NativeVlan = DefaultVLAN
-		pc.TrunkVlans = DefaultVLANRange
-		pc.UserCfgdFlags = "admin_layer,admin_state"
 		if req.Interface.Spec.AdminState == v1alpha1.AdminStateUp {
 			pc.AdminSt = AdminStUp
 		}
+		// Note: Layer 3 port-channel interfaces are not yet supported
+		pc.Layer = Layer2
+		pc.Mode = SwitchportModeAccess
+		pc.AccessVlan = DefaultVLAN
+		pc.NativeVlan = DefaultVLAN
+		pc.TrunkVlans = DefaultVLANRange
+		pc.UserCfgdFlags = "admin_state"
+
+		pc.MTU = DefaultMTU
+		if req.Interface.Spec.MTU != 0 {
+			pc.MTU = req.Interface.Spec.MTU
+			pc.UserCfgdFlags = "admin_mtu," + pc.UserCfgdFlags
+		}
+
+		pc.PcMode = PortChannelModeActive
+		switch m := req.Interface.Spec.Aggregation.ControlProtocol.Mode; m {
+		case v1alpha1.LACPModeActive:
+			pc.PcMode = PortChannelModeActive
+		case v1alpha1.LACPModePassive:
+			pc.PcMode = PortChannelModePassive
+		default:
+			return fmt.Errorf("iface: unknown LACP mode: %s", m)
+		}
+
 		if req.Interface.Spec.Switchport != nil {
 			switch req.Interface.Spec.Switchport.Mode {
 			case v1alpha1.SwitchportModeAccess:
@@ -590,14 +606,34 @@ func (p *Provider) EnsureInterface(ctx context.Context, req *provider.InterfaceR
 			}
 		}
 
-		// for _, ifc := range req.Interface.Spec.PhysicalInterfaces {
-		// 	name, err := ShortName(ifc.Name)
-		// 	if err != nil {
-		// 		return provider.Result{}, err
-		// 	}
-		//  tdn := "System/intf-items/phys-items/PhysIf-list[id=" + name + "]"
-		//  pc.RsmbrIfsItems.RsMbrIfsList = append(pc.RsmbrIfsItems.RsMbrIfsList, &PortChannelMember{TDn: tdn})
-		// }
+		for _, member := range req.Members {
+			n, err := ShortNamePhysicalInterface(member.Spec.Name)
+			if err != nil {
+				return err
+			}
+			pc.RsmbrIfsItems.RsMbrIfsList = append(pc.RsmbrIfsItems.RsMbrIfsList, NewPortChannelMember(n))
+		}
+
+		v := new(VPCIfItems)
+		if err := p.client.GetConfig(ctx, v); err != nil && !errors.Is(err, gnmiext.ErrNil) {
+			return err
+		}
+
+		// Delete the existing VPC interface entry if the MultiChassisID has changed or got removed.
+		if vpc := v.GetListItemByInterface(name); vpc != nil {
+			if req.MultiChassisID == nil || int(*req.MultiChassisID) != vpc.ID {
+				if err := p.client.Delete(ctx, vpc); err != nil {
+					return err
+				}
+			}
+		}
+
+		if req.MultiChassisID != nil {
+			v := new(VPCIf)
+			v.ID = int(*req.MultiChassisID)
+			v.SetPortChannel(name)
+			conf = append(conf, v)
+		}
 
 		conf = append(conf, pc)
 
@@ -622,10 +658,14 @@ func (p *Provider) DeleteInterface(ctx context.Context, req *provider.InterfaceR
 		return err
 	}
 
-	addr := new(AddrItem)
-	addr.ID = name
+	conf := make([]gnmiext.Configurable, 0, 3)
+	if req.Interface.Spec.Type != v1alpha1.InterfaceTypeAggregate {
+		addr := new(AddrItem)
+		addr.ID = name
 
-	conf := []gnmiext.Configurable{addr}
+		conf = append(conf, addr)
+	}
+
 	switch req.Interface.Spec.Type {
 	case v1alpha1.InterfaceTypePhysical:
 		p := new(PhysIf)
@@ -635,10 +675,27 @@ func (p *Provider) DeleteInterface(ctx context.Context, req *provider.InterfaceR
 		stp := new(SpanningTree)
 		stp.IfName = name
 		conf = append(conf, stp)
+
 	case v1alpha1.InterfaceTypeLoopback:
 		lb := new(Loopback)
 		lb.ID = name
 		conf = append(conf, lb)
+
+	case v1alpha1.InterfaceTypeAggregate:
+		pc := new(PortChannel)
+		pc.ID = name
+		conf = append(conf, pc)
+
+		v := new(VPCIfItems)
+		if err := p.client.GetConfig(ctx, v); err != nil && !errors.Is(err, gnmiext.ErrNil) {
+			return err
+		}
+
+		// Make sure to delete any associated VPC interface.
+		if vpc := v.GetListItemByInterface(name); vpc != nil {
+			conf = append(conf, vpc)
+		}
+
 	default:
 		return fmt.Errorf("unsupported interface type: %s", req.Interface.Spec.Type)
 	}
@@ -647,11 +704,16 @@ func (p *Provider) DeleteInterface(ctx context.Context, req *provider.InterfaceR
 }
 
 func (p *Provider) GetInterfaceStatus(ctx context.Context, req *provider.InterfaceRequest) (provider.InterfaceStatus, error) {
+	name, err := ShortName(req.Interface.Spec.Name)
+	if err != nil {
+		return provider.InterfaceStatus{}, err
+	}
+
 	var operSt OperSt
 	switch req.Interface.Spec.Type {
 	case v1alpha1.InterfaceTypePhysical:
 		phys := new(PhysIfOperItems)
-		phys.ID = req.Interface.Spec.Name
+		phys.ID = name
 		if err := p.client.GetState(ctx, phys); err != nil && !errors.Is(err, gnmiext.ErrNil) {
 			return provider.InterfaceStatus{}, err
 		}
@@ -659,11 +721,19 @@ func (p *Provider) GetInterfaceStatus(ctx context.Context, req *provider.Interfa
 
 	case v1alpha1.InterfaceTypeLoopback:
 		lb := new(LoopbackOperItems)
-		lb.ID = req.Interface.Spec.Name
+		lb.ID = name
 		if err := p.client.GetState(ctx, lb); err != nil && !errors.Is(err, gnmiext.ErrNil) {
 			return provider.InterfaceStatus{}, err
 		}
 		operSt = lb.OperSt
+
+	case v1alpha1.InterfaceTypeAggregate:
+		pc := new(PortChannelOperItems)
+		pc.ID = name
+		if err := p.client.GetState(ctx, pc); err != nil && !errors.Is(err, gnmiext.ErrNil) {
+			return provider.InterfaceStatus{}, err
+		}
+		operSt = pc.OperSt
 
 	default:
 		return provider.InterfaceStatus{}, fmt.Errorf("unsupported interface type: %s", req.Interface.Spec.Type)
